+++ core/slave/index.js
@@ -10,6 +10,7 @@ exports.middleware = function *(next) {
   if (this.url === '/' && this.method === 'POST') {
     if (process._isReady) {
       const post = yield _.parse(this);
+
       process.send({
         message: 'bindSlave',
         data: post
@@ -29,17 +30,23 @@ exports.init = function() {
   if (cluster.isMaster) {
     process.slaveManager = process.slaveManager || new Manager();
 
-    Object.keys(cluster.workers).forEach((id) => {
-      cluster.workers[id].on('message', (e) => {
+    Object.keys(cluster.workers).forEach(id => {
+      cluster.workers[id].on('message', e => {
         switch (e.message) {
           case 'bindSlave':
-            process.slaveManager.bind(e.data);
+            const data = e.data;
+            if (data.supportiOS !== undefined) {
+              data.supportiOS = data.supportiOS === 'true';
+            }
+
+            process.slaveManager.bind(data);
             break;
-          case 'dispatch':
+          case 'dispatchTask':
             process.slaveManager.dispatch(e.data);
             break;
         }
       });
+
       cluster.workers[id].send({
         message: 'slaveReady'
       });