+++ core/task/check.js
@@ -10,23 +10,23 @@ const Task = models.Task;
 const TIMEOUT = 60; // minutes
 
 module.exports = co.wrap(function *() {
-  const task = new Task();
-  const outDatedTasks = yield task.getOutdatedTasks(TIMEOUT);
-  let taskQueue = [];
-
-  outDatedTasks.forEach(data => {
-    logger.debug('Task %s is outdated, put it back to task queue.', data._id);
-    _.merge(data, {
-      status: 0,
-      start_at: null,
-      extra: '',
-      slaveId: '',
-      result: []
-    });
-    taskQueue.push(task.updateById(data._id, data));
-  });
-
   try {
+    const task = new Task();
+    const outDatedTasks = yield task.getOutdatedTasks(TIMEOUT);
+    let taskQueue = [];
+
+    outDatedTasks.forEach(data => {
+      logger.debug('Task %s is outdated, put it back to task queue.', data._id);
+      _.merge(data, {
+        status: 0,
+        start_at: null,
+        extra: '',
+        slaveId: '',
+        result: []
+      });
+      taskQueue.push(task.updateById(data._id, data));
+    });
+
     yield taskQueue;
   } catch(e) {
     logger.warn(e);