+++ init/worker.js
@@ -16,3 +16,7 @@ if (cluster.isWorker) {
     }
   });
 }
+
+process.on('uncaughtException', function(err) {
+  console.error('Worker Error caught in uncaughtException event:', err.stack);
+});