+++ web/index.js
@@ -1,13 +1,22 @@
 'use strict';
 
-const koa = require('koa');
+require('babel/register')({
+  extensions: ['.jsx']
+});
+
 const EOL = require('os').EOL;
+const koa = require('koa');
+const plugin = require('reliable-plugin');
 
 const router = require('./router');
-const plugins = require('../core/plugin');
 const _ = require('../common/utils/helper');
 const middlewares = require('./middlewares');
+const auth = require('./middlewares/auth');
 const logger = require('../common/utils/logger');
+const layout = require('./views/common/layout');
+const model = require('../common/models/');
+const pluginModel = model.Plugin;
+const taskModel = model.Task;
 
 exports.init = (options, callback) => {
 
@@ -19,7 +28,12 @@ exports.init = (options, callback) => {
 
   middlewares(app);
 
-  plugins.register(app);
+  plugin(app, {
+    pluginModel,
+    taskModel,
+    layout,
+    auth
+  });
 
   router(app);