+++ web/middlewares/i18n.js
@@ -3,18 +3,26 @@
 const i18n = require('../resources/i18n');
 const _ = require('../../common/utils/helper');
 
-const locales = ['zh_CN', 'en_US'];
+const locales = ['zh-CN', 'en-US'];
 
 module.exports = function(app) {
   return function *inject(next) {
-    const locale = this.query.locale || this.cookies.get('locale') || app._options.site.locale;
-    if (locale) {
-      if (_.includes(locales, locale)) {
-        app.options['locale'] = locale;
-        this.gettext = i18n(locale);
-        this.cookies.set('locale', locale);
-      }
+    let locale = this.query.locale || this.cookies.get('locale') || this.request.acceptsLanguages();
+
+    if (typeof locale === 'string') {
+      locale = [locale];
     }
+
+    // zh_CN -> zh-CN
+    locale = locale.map(l => l.replace(/_/, '-'));
+
+    locale = _.intersection(locale, locales)[0] || app._options.site.locale;
+    app._options.locale = locale;
+    const i = i18n(locale);
+    app.addPluginText = i.addPluginText;
+    this.gettext = i.gettext;
+    this.cookies.set('locale', locale);
+
     yield next;
   };
 };