+++ web/views/logFormatter/index.js
@@ -1,13 +1,14 @@
 'use strict';
 
-var EOL = require('os').EOL;
-var Convert = require('ansi-to-html');
+const EOL = require('os').EOL;
+const Convert = require('ansi-to-html');
 
-var screenshot = require('./screenshot');
-var performance = require('./performance');
+const screenshot = require('./screenshot');
+const performance = require('./performance');
+const trTpl = require('./tr-tpl');
 
-var perfFragments;
-var convert = new Convert();
+let perfFragments;
+const convert = new Convert();
 
 function handler(type, log) {
   if (typeof type !== 'string') {
@@ -35,24 +36,52 @@ function handler(type, log) {
 }
 
 function hasPerformanceData(logs) {
-  var perfPattern = /\<\# performance[\s\S]*\#\>/g;
+  const perfPattern = /\<\# performance[\s\S]*\#\>/g;
   return perfPattern.test(logs);
 }
 
+function tableWrap(raw) {
+  let i = 0;
+
+  return raw.split('\n').map(v => {
+    if (v.trim()) {
+      if (v.indexOf('reliable_timing_start') > -1) {
+        return trTpl.timingStart();
+      }
+
+      if (v.indexOf('reliable_timing_end') > -1) {
+        return trTpl.timingEnd(v);
+      }
+
+      if (v.indexOf('reliable_fold:start') > -1) {
+        return trTpl.foldStart(v);
+      }
+
+      if (v.indexOf('reliable_fold:end') > -1) {
+        return trTpl.foldEnd(v);
+      }
+
+      i++;
+      return trTpl.defaultTpl(v, i);
+    }
+  });
+}
+
 function beautify(logs) {
-  return convert.toHtml(logs).replace(/\n+/g, '\n\n');
+  const raw = convert.toHtml(safe(logs)).replace(/\n+/g, '\n\n');
+
+  const table = tableWrap(raw);
+  return table.join('');
 }
 
 function format(logs) {
   logs = beautify(logs);
-  var pattern = /\<\#([\s\S]*?)\#\>/g;
+  const pattern = /\<\#([\s\S]*?)\#\>/g;
   perfFragments = hasPerformanceData(logs) ? EOL + '<p>Performance Table:</p>' : '';
   return (logs.replace(pattern, function(matchStr, identifyStr) {
-    var arr = identifyStr.split('|');
-    var type = arr[0];
-    var log = arr[1];
+    const [type, log] = identifyStr.split('|');
     if (type && log) {
-      var res = handler(type, log);
+      const res = handler(type, log);
       if (res) {
         return res.trim();
       }
@@ -61,4 +90,10 @@ function format(logs) {
   }) + perfFragments).replace(/\n{2,}/g, '\n\n');
 }
 
-module.exports = format;
+function safe(log) {
+  return log.replace(/</g, '&lt;').replace(/>/g, '&gt;');
+}
+
+exports.format = format;
+exports.safe = safe;
+