+++ HISTORY.md
@@ -1,7 +1,31 @@
-# 1.0.1  / 2016-03-24
+# 2.0.0  / 2016-12-19
+* feat: better task log, support line number & log folding;
 
-  * Tweak Badgeboard to Support Title
+# 1.1.2  / 2016-11-08
+* refactor: passing Task model to plugin;
+* chore: add missing try/catch.
+
+# 1.1.1  / 2016-11-02
+* fix: init redis slave field when bootstrap master;
+* fix: don't display unregistered plugins.
+
+# 1.1.0  / 2016-10-31
+* feat: marking slave whether support iOS task;
+* fix: subscribe email bug;
+* fix: i18n bug.
+
+# 1.0.8  / 2016-10-27
+* feat: support plugin;
+* fix: badgeboard bug.
+
+# 1.0.4  / 2016-08-29
+* fix: i18n bug.
+
+# 1.0.3  / 2016-08-18
+* refactor: better i18n.
+
+# 1.0.1  / 2016-03-24
+* refactor: tweak badgeboard to support title.
 
 # 1.0.0  / 2016-03-16
-
-  * First Release
+* First Release.